# openapi.yaml
openapi: 3.0.0
info:
  title: API de Superhéroes
  description: API para gestionar superhéroes y sus mascotas con sistema de cuidado y autenticación JWT
  version: 1.0.0
  contact:
    name: VaneSolis
    url: https://github.com/VaneSolis/SuperHeroesAPI
servers:
  - url: http://localhost:3001
    description: Servidor de desarrollo
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: Nombre de usuario
        password:
          type: string
          description: Contraseña
    RegisterRequest:
      type: object
      required:
        - username
        - password
        - email
      properties:
        username:
          type: string
          minLength: 3
          description: Nombre de usuario (mínimo 3 caracteres)
        password:
          type: string
          minLength: 6
          description: Contraseña (mínimo 6 caracteres)
        email:
          type: string
          format: email
          description: Email válido
    AuthResponse:
      type: object
      properties:
        message:
          type: string
        token:
          type: string
          description: JWT token
        user:
          type: object
          properties:
            id:
              type: integer
            username:
              type: string
            role:
              type: string
paths:
...
  /auth/login:
    post:
      summary: Iniciar sesión
      description: Autenticar usuario y obtener token JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login exitoso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Credenciales inválidas
        '400':
          description: Datos de entrada inválidos
  /auth/register:
    post:
      summary: Registrar nuevo usuario
      description: Crear una nueva cuenta de usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Usuario registrado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Datos de entrada inválidos o usuario ya existe
  /auth/verify:
    get:
      summary: Verificar token JWT
      description: Verificar si el token JWT es válido
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Token válido
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  valid:
                    type: boolean
                  user:
                    type: object
        '401':
          description: Token no proporcionado
        '403':
          description: Token inválido
...